:rocket: finished implementing secure login for words api

+++ server/routes/words.js
@@ -17,14 +17,19 @@ const wordModel = require('../models/word.js')
 // ============================
 wordsRouter.post('/', async (req, res) => {
   //TODO: Expand Error codes (Specifically error code for duplicate values)
-  const Word = new wordModel(req.body)
+  console.log(req.user.roles)
+  if (req.user.roles.includes('editor')) {
+    const Word = new wordModel(req.body)
 
-  try {
-    await Word.save()
-    res.status(200).send(Word)
-  } catch (err) {
-    console.error(err)
-    res.status(500).send(err)
+    try {
+      await Word.save()
+      res.status(200).send(Word)
+    } catch (err) {
+      console.error(err)
+      res.status(500).send(err)
+    }
+  } else {
+    res.sendStatus(403)
   }
 })
 // ==================
@@ -44,31 +49,39 @@ wordsRouter.get('/', async (req, res) => {
 // UPDATE
 // ==================
 wordsRouter.patch('/:_id', async (req, res) => {
-  try {
-    // console.log(req.body)
-    const Word = await wordModel.findOneAndUpdate(
-      { _id: req.params._id },
-      req.body
-    )
-    // await wordModel.save()
-    res.status(200).send(Word)
-  } catch (err) {
-    console.log(err)
-    res.status(500).send(err)
+  if (req.user.roles.includes('editor')) {
+    try {
+      // console.log(req.body)
+      const Word = await wordModel.findOneAndUpdate(
+        { _id: req.params._id },
+        req.body
+      )
+      // await wordModel.save()
+      res.status(200).send(Word)
+    } catch (err) {
+      console.log(err)
+      res.status(500).send(err)
+    }
+  } else {
+    res.sendStatus(403)
   }
 })
 // =================
 // DELETE
 // =================
 wordsRouter.delete('/:_id', async (req, res) => {
-  try {
-    await wordModel.findOneAndDelete({ _id: req.params._id })
+  if (req.user.roles.includes('editor')) {
+    try {
+      await wordModel.findOneAndDelete({ _id: req.params._id })
 
-    // if (!Words) res.status(404).send('No item found')
-    res.status(200).send()
-  } catch (err) {
-    console.error(err)
-    res.status(500).send(err)
+      // if (!Words) res.status(404).send('No item found')
+      res.status(200).send()
+    } catch (err) {
+      console.error(err)
+      res.status(500).send(err)
+    }
+  } else {
+    res.sendStatus(403)
   }
 })
